Additional command line arguments, call docker from node
The node wrapper script has a flag that causes it to try to invoke
the docker container. Right now, it uses a hardcoded image id. If
the flag is not set, it tries to execute the validator binary directly.

The validator binary has some additional command line arguments. To
support this, the TCLAP library is added.

Temporarily changing the Dockerfile to use ubuntu until I figure out
what nonsense is going on with alpine.