Schema version is required argument
When calling the script, the two required arguments are the path
to the data file and the schema version. If the specified version
is not found, the available versions are listed.

Remove option for using the docker container. Always use the docker
container. Remove function for running validator executable directly.