Support xml validation
XML files can now be validated. The validator will attempt to use
the extension on the data file to determine whether to use XML or
JSON validation. An option is provided to specify the data format
explicitly. The XML validator is implemented in Java, and has a
slightly different format for its command-line arguments. The
Dockerfile is updated to contain the XML validator and an entry
point script to determine which validator executable to run.