Allow user to validate multiple files within a ZIP
After validating a file within a ZIP archive, ask the user if they
want to validate another file in the ZIP. If they say yes, show the
file selection menu again. If they say no, exit. The user is allowed
to validate another file regardless of the validation result, as long
as the validator container executes successfully.