Basic support for processing referenced files in table of contents
This is the most basic version of support for this feature to provide
proof of concept. When a local file is valid against the table of
contents schema, open it and collect the URLs of referenced files.
Ask the user if they want to validate those files. If they say yes,
download and validate the files against the appropriate schemas. If
an output file is specified, the outputs for the referenced files are
appended to that file.