Get referenced files while validating
The validator executable creates a JSON file with lists of referenced
JSON files. The types of files referenced depends on the type of
file being validated.

Note that this is only a partial implementation, applicable to
validation of local files. Additionally, files found with references
will not have their references checked.