Reorganize functions for working with discovered files
A table of contents file can have in-network and allowed amounts files.
An in-network file can have provider reference files. As long as we
start with only one file (either local or downloaded), we can try
exploring discovered files.

SchemaManager now keeps track of whether it is in strict validation
mode, rather than requiring the strict parameter to be passed to its
various functions.

DockerManager is currently just a container for two docker-related
functions, but will eventually also be responsible for storing some
values to reduce the amount of parameters that have to be sent to
each function call.

Eventually, there will be some kind of implementation for supporting
discovered files when working with a ZIP archive. But, this will likely
require a more complex UI for choosing between files in the ZIP archive
and discovered files.