Fix validator path tracking bug
The validator previously would track its traversal path incorrectly
when an array contained objects. This fix allows the validator to
correctly find provider references.

Add a debug flag that provides output of the startup command and
docker container run commands.

SchemaManager returns null when asked for a schema that doesn't exist.

Add tests for SchemaManager and DockerManager. Remove functions and
tests for utils for the features that are implemented by those two
classes. Update tests for commands to use SchemaManager and
DockerManager instead of removed utils functions.