Add tests for SchemaManager.ensureRepo
These tests require mocking the exec function, since we don't want
these tests to try to actually clone anything. This makes things a
little bit complicated, because other tests specifically do not want
to mock this function. Mock the module, but add specific mock
implementations in cases where we want to call the actual version
of the exec function.