Add option to always confirm downloads
The -y option can be provided to automatically respond "yes" to
any user prompt that asks if you want to download a specific file.
These prompts appear when the size of the file exceeds a threshold
and when the size of the file is unknown.

Download functions moved to DownloadManager class. This class stores
the value of the "yes to all downloads" option, which allows it to
skip showing prompts when set.