Additional data collection for environmental scan
This is an early implementation that just prints some additional
info to stdout. The full implementation will store the information
in an output file. The additional information tracked currently includes:
- number of negotiated_price objects for an in-network-rates file
- number and location of additional_info properties in negotiated_price objects

Output path is now a directory, since multiple output files are produced.
Add "fail-fast" flag to end validation on the first error. It may
be useful in some cases to continue validation so that more information
can be collected.