Stream output of reported items
Instead of collecting items that match report paths, immediately write
them to the report file. All paths are now assumed to be wildcard paths,
so the "*" that would sometimes be at the end is removed. This also makes
it easier to report values that are objects or arrays.

Path counting is temporarily removed until I make sure it works.